+++ src/main/java/com/jbialy/rce/utils/OutOfRangeException.java
@@ -1,7 +1,7 @@
 package com.jbialy.rce.utils;
 
 public class OutOfRangeException extends RuntimeException {
-    public OutOfRangeException(int desiredIndex, int i, int i1) {
+    public OutOfRangeException(int desiredIndex, long i, long i1) {
 
     } //todo
 }
